// frontend/src/enums/ProgrammingLanguage.tsx
export enum LanguageVersion {
  PYTHON_3_8 = "PYTHON_3_8",
  PYTHON_2_7 = "PYTHON_2_7",

  C_GCC_7_4 = "C_GCC_7_4",
  C_GCC_8_3 = "C_GCC_8_3",
  C_GCC_9_2 = "C_GCC_9_2",
  C_CLANG_7_0 = "C_CLANG_7_0",

  CPP_GCC_7_4 = "CPP_GCC_7_4",
  CPP_GCC_8_3 = "CPP_GCC_8_3",
  CPP_GCC_9_2 = "CPP_GCC_9_2",
  CPP_CLANG_7_0 = "CPP_CLANG_7_0",

  JAVA_OPENJDK_13 = "JAVA_OPENJDK_13",
  JAVA_OPENJDK_11 = "JAVA_OPENJDK_11",

  JAVASCRIPT_NODEJS_12 = "JAVASCRIPT_NODEJS_12",

  CSHARP_MONO = "CSHARP_MONO",

  GO = "GO",
  KOTLIN = "KOTLIN",
  RUST = "RUST",
  RUBY = "RUBY",
  PHP = "PHP",
  LUA = "LUA",
  SWIFT = "SWIFT",
  R = "R",

  BASH = "BASH",
  BASIC = "BASIC",
  CLOJURE = "CLOJURE",
  COBOL = "COBOL",
  COMMON_LISP = "COMMON_LISP",
  D = "D",
  ELIXIR = "ELIXIR",
  ERLANG = "ERLANG",
  FORTRAN = "FORTRAN",
  GROOVY = "GROOVY",
  HASKELL = "HASKELL",
  OCAML = "OCAML",
  OCTAVE = "OCTAVE",
  PASCAL = "PASCAL",
  PERL = "PERL",
  PROLOG = "PROLOG",
  OBJECTIVEC = "OBJECTIVEC",
  FSHARP = "FSHARP",
  SQL = "SQL",
  VISUAL_BASIC = "VISUAL_BASIC",
  MULTIFILE_PROGRAM = "MULTIFILE_PROGRAM",
  EXECUTABLE = "EXECUTABLE",
}